Update the create corporate docs with current state of DTOs
Seems like some of those fields are not required and some we don't even use but this is the current state and until we change the annotations those fields are mandatory.
I've POSTed that json and it seems to be working.

This still need some major update but this version is better then what's currently deployed so shipping this.